% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportNifti.R
\name{exportNifti}
\alias{exportNifti}
\title{Export data in Nifti format}
\usage{
exportNifti(ims3d, xyz_um = c(20, 20, 20), name_prefix = "my3d_data")
}
\arguments{
\item{ims3d}{combined, processed 3D IMS dataset}

\item{xyz_um}{spatial resolution information in microns for all three dimensions}

\item{name_prefix}{character string to be prepended to exported nifty image filenames}
}
\description{
Exports reconstructed 3D arrays which can be opened and visualized in FIJI ImageJ, Slicer or other scientific imaging software. Function writes *.nii files to the current working directory.
}
