% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get3DTemplateImgs.R
\name{get3DTemplateImgs}
\alias{get3DTemplateImgs}
\title{Get 3D template images for registration}
\usage{
get3DTemplateImgs(ims, canvas_xy = c(700, 700), column = "PC1")
}
\arguments{
\item{ims}{Cardinal data of class: MSImagingExperiment, PCA2, SpatialShrunkenCentroids2}

\item{canvas_xy}{standardized canvas size for all images, must accomodate all IMS pixels}

\item{column}{name of column to extract if character, or m/z value if numeric}
}
\value{
named list of images
}
\description{
Generate a named list  of template images based on Cardinal IMS or analysis data (m/z images, PCA, Spatial Shrunken Centroids clustering)
}
