% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge3DIMS.R
\name{merge3DIMS}
\alias{merge3DIMS}
\title{Merge a list of 2D MSImagingExperiments to 3D}
\usage{
merge3DIMS(ims3d)
}
\arguments{
\item{ims3d}{list of processed 3D IMS dataset}
}
\value{
3D MSImagingExperiment with z-dimension added
}
\description{
Merges MSImagingExperiments to 3D using the 'run' variable of each
}
