% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg3DIMS.R
\name{reg3DIMS}
\alias{reg3DIMS}
\title{Register serial 3D IMS data}
\usage{
reg3DIMS(reg_images, mid_point = 5, flip_positions = c(),
  flip_type = c())
}
\arguments{
\item{reg_images}{list of 2D matrices representing template images}

\item{mid_point}{depth from which registration should proceed}

\item{flip_type}{optional, 1 (horizontral) or 2(vertical), length must match length of flip_position}

\item{flip_position}{optional, depth at which the image is mirror, length must match length of flip_type}
}
\value{
named list of NiftyReg transformations per 2D section
}
\description{
This function registers a sequence of 2D images saved as a list of 2D matrices into a 3D reconstruction, returning a list of transformations
}
