% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform3DIMS.R
\name{transform3DIMS}
\alias{transform3DIMS}
\title{Transfom 2D IMS data into 3D reconstruction}
\usage{
transform3DIMS(ims, tforms)
}
\arguments{
\item{ims}{Cardinal data of class: MSImagingExperiment}

\item{tforms}{named list of NiftyReg transformations per 2D section, generated by reg3DIMS function}
}
\value{
alphanumerically ordered named list of MSImagingExperiment's transformed to their 3D coordinates
}
\description{
Generates a name ordered list of MSImagingExperiment data that has been transformed into 3D
}
